some fix

+++ src/main/java/iw/gdupo/restaurant/dto/MenuDTO.java
@@ -2,6 +2,8 @@ package iw.gdupo.restaurant.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class MenuDTO {
 
@@ -10,5 +12,5 @@ public class MenuDTO {
     private String shortDescription;
     private String description;
     private Integer weight;
-    private Integer price;
+    private BigDecimal price;
 }
